requirements.txt
numpy
pandas
librosa
soundfile
scikit-learn
joblib
pyyaml
matplotlib
seaborn
